chore: Atualizar config, remover testes, refatorar posts
Aqui está a tradução para português:

- Atualizar a configuração do package.json: alterar o rootDir, adicionar moduleNameMapper e ajustar o coverageDirectory.
- Removidos os testes de AuthController do arquivo auth.controller.spec.ts.
- Removido o módulo de testes do serviço de autenticação e os testes relacionados.
- Adicionar um novo arquivo de teste para o serviço de postagens com testes para encontrar postagens, rascunhos e associações, criar e atualizar postagens, e excluir postagens.
- Removidos os testes para UserSchoolAssociationController no arquivo association.controller.spec.ts.
- Atualizados os testes para o serviço de postagens: Removido arquivo de teste desatualizado, refatorados os testes para a recuperação de postagens, rascunhos, postagens associadas, criação e atualização de postagens.
