+++ package.json
@@ -79,7 +79,10 @@
       "json",
       "ts"
     ],
-    "rootDir": "src",
+    "rootDir": "./src",
+    "moduleNameMapper": {
+      "@/(.*)$": "<rootDir>/$1"
+    },
     "testRegex": ".*\\.spec\\.ts$",
     "transform": {
       "^.+\\.(t|j)s$": "ts-jest"
@@ -87,7 +90,7 @@
     "collectCoverageFrom": [
       "**/*.(t|j)s"
     ],
-    "coverageDirectory": "../coverage",
+    "coverageDirectory": "./coverage",
     "testEnvironment": "node"
   }
 }
